+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## 0.1.7
 
+- Changed Widget to use ConfigKey type instead of u8, as this might change in a later release.
 
 ## 0.1.6